requirements.txt
Django==4.2.28
django-crispy-forms==2.5.0
django-allauth[socialaccount]==65.14.0
pyjwt==2.11.0
cryptography==46.0.4
google-api-python-client==2.187.0
djangorestframework==3.16.1
django-rest-swagger==2.2.0
coreapi==2.3.3
pyyaml==6.0.3
uritemplate==4.2.0
drf-yasg==1.21.14
python-dotenv==1.2.1
requests==2.32.5
django-autocomplete-light==3.12.1
pymongo==4.10.1
